# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rexxgo API",
        "description": "API documentation for Auth and Wallet modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Base server"
        }
    ],
    "paths": {
...
        "/api/wallets/{accountNumber}/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet balance",
                "operationId": "cd7f4a6fbf43e9bf5420dc7b89c1f9da",
                "parameters": [
                    {
                        "name": "accountNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}